Remove the duplicated DropPath block (We already had it). Added some comments
